Refactors types and cleans up files
