+++ src/index.ts
@@ -2,8 +2,13 @@ export { SignIn as SignInWithFractal } from 'components/sign-in';
 export { UserContextProvider as SignInWithFractalProvider } from 'context/user';
 export type { SignInProps as SignInWithFractalProps } from 'components/sign-in';
 
-export { Scope as FractalSdkScope } from 'types/scope';
-export type { FractalUser } from 'types/user';
+export type {
+  FractalUser,
+  FractalUserWallet,
+  FractalCoin,
+  FractalItem,
+  Scope as FractalSdkScope,
+} from 'types';
 
 export { useFractalUser } from 'hooks/use-fractal-user';
 export { useFractalUserWallet } from 'hooks/use-fractal-user-wallet';
