+++ src/types/coin.ts
@@ -1,0 +1,9 @@
+export interface FractalCoin {
+  address: string;
+  amount: number;
+  decimals: number;
+  logoUri: string;
+  name: string;
+  symbol: string;
+  uiAmount: string;
+}
